refactor: set up gormigrate
